// app-turnOver-v1/src/app/models/FitCultural.ts
export interface FitCultural {
  nome: string;
  data: string;
  trabalhoEquipe?: boolean;
  adaptabilidade: boolean;
  comunicativo: boolean;
  pensamentoCritico?: boolean;
  gestaoTempo?: boolean;
  empatia?: boolean;
  alinhadoValores: boolean;
  proatividade?: boolean;
  iniciativa: boolean;
  observacoes?: string;

}